requirements.txt
RPi.GPIO
spidev
soco-cli
pyserial
